Setup a 2nd cell for the subnode in multinode setups
This creates a nova_subnode_cell database and subnode-cell
cells v2 cell. This is done when setting up the nova,
nova_api and nova_cell0 databases on the primary node.

We have to move the call to 'nova-manage db sync' so that
it runs after the subnode cell is created so that it's database
schema is migrated along with the main (nova) and nova_cell0
databases.

This change is necessary for testing out a multi-cell deployment
with cells v2.

Change-Id: I6d263c81579f8c52dcf4cb9a0085c64545ee13e5
Depends-On: I4fb1c966e6e011b73d42a1629718abf1adc8125f