Set MYSQL_HOST to SERVICE_HOST if available
The multinode setup in devstack-gate sets the SERVICE_HOST
variable to be the IP of the primary node. The SERVICE_HOST
variable isn't set for non-multinode jobs.

When SERVICE_HOST isn't used in a multinode job to set the
MYSQL_HOST value, it defaults to 127.0.0.1, which is fine
for anything on the primary node needing to connect to the
local databases.

On the subnode, devstack-gate sets the DATABASE_HOST variable
in localrc which the subnode uses to configure a database
connection back to the primary node. That's all fine.

The problem is with cells v2 we're setting up the cells on
the primary node and the database connection URL is stored
in the nova_api database, and right now that's pointing at
127.0.0.1 without this fix. When that happens, the subnode
can't connect to the database in the cell because there isn't
a database local to the subnode.

So this fix simply uses the SERVICE_HOST variable, if set,
to determine the MYSQL_HOST variable which is used if
DATABASE_HOST isn't set in localrc.

Change-Id: I799587d1019cf81ff6d667f479f876f3e5962943