Update patch set 1
Patch Set 1: Code-Review-1

I might be getting lost to the discussion, but why this selector is not being done in the test itself (volume-feature-enabled)? And then the job that test it should add to the job config? That seems the sane way to do it, and is the same way other optional features like backup, clone and snapshot where done. So we should have a 'manage_snapshot' option as well. Then add it to the LVM job.

Patch-set: 1
Label: Code-Review=-1