Installin pip3 when installing python3
The install_python3 function installs python3 and python3-dev
packages, but it does not install python3-pip. However function
install_devstack_tools uses that function to install python3
in the system, and then it relies on using pip3 for installing
devstack-tools however pip3 might not be there since it just
installed python3.
This commit adds the installation of python3-pip as part of the
install_python3 function.

Change-Id: I0915354b4bc4f598ec7310a7978b604e93f6cc5a
Closes-Bug: 1659083