Conflict adding cinder type (lvmdriver-1)
In a multinode environment, c-vol should be enabled on the second node
which triggers volume type creation (as per documentation). When running
stack.sh on the second node,an error is reported cinder is trying to
create a volume type that already exists.

In this fix, check to see if the volume type exists before creating it.

Closes-Bug: #1390028

Change-Id: Id7ae00456f1be5c1aed9461379a99832f8faf5ed
Signed-off-by: Amy Fong <amy.fong@windriver.com>