Update patch set 3
Abandoned

I don't think we need this anymore.

> Because we're calling discover_hosts from the subnode in devstack-gate

That's actually not correct, we run discover_hosts from the primary node in the d-g change:

https://review.openstack.org/#/c/424252/4/devstack-vm-gate.sh@238

And I *think* I just needed this change to get this working:

https://review.openstack.org/#/c/420976/

Which was trying to setup a 2nd cell on the subnode, which isn't going to work right now, and we're too tight for it to make FF anyway, so let's just abandon this for now.

Patch-set: 3
Status: abandoned