Update patch set 1
Patch Set 1:

You're just looking at the wrong file. It's in the human readable dstat file:

http://logs.openstack.org/20/428220/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/6d6c703/logs/screen-dstat.txt.gz

The reason we have different arguments for dstat is because there are bugs in dstat which weren't fixed in the version of dstat installed on the system. The top fields were consistently crashing dstat so we decided not to put them in csv. The csv file is there for the stackviz output from the run:

http://logs.openstack.org/20/428220/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/6d6c703/logs/stackviz/#/stdin/timeline

Although it looks like there is a bug in that so the line graphs for system load aren't being used.

Patch-set: 1