Added memory_tracker service
This service lists processes that lock memory pages from swapping to
disk. The service may be helpful when debugging oom-killer job failures
in gate in case when dstat shows that swap is not fully used when
oom-killer is triggered.

The service may be extended with other info in the future.

Change-Id: I1dca120448ee87930fe903fd81277b58efaefc92