Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

-1 for slight issue with the loop as stated

i wonder if maybe it's worth just dumping the whole summary line each time anyway, as a detailed mem snapshot every 30s or so.  would be annoying if we find we want another stat to have to re-write this again

that would make it similar to the peakmem tracker i guess, which i put in for similar debugging.  the difference being that only logs when it sees the memory going down.  i like the loop and the pmap tool, i think that's probably a better version of what i was going for with the "ps" output too

Patch-set: 1
Label: Code-Review=-1