Differentiate between DEFAULT_IMAGE_NAME and filename
The DEFAULT_IMAGE_NAME variable is used to reference the name of the
default image in glance after it has been uploaded by devstack. It is
used both inside and outside of devstack for that purpose. However, when
configuring tempest there are some tests which also do image uploads and
need a filename for specifying which file they should upload into glance
for testing purposes. Previously we were just using DEFAULT_IMAGE_NAME
for both purposes, but this causes a conflict if the name of the image
we upload into glance does not have a file extension. So instead of
conflating the things this commit differentiates between them and adds a
new DEFAULT_IMAGE_FILE_NAME variable to use for this purpose.

This conflicts because there were several intermediate changes before
this was finished. This just skips to the end state. Additionally the
change id is removed because to get things working in the gate
consistently this needs to land in a very specific order. This can't be
done if the change-ids all match.

 Conflicts:
	lib/tempest
	stackrc

(cherry picked from commit befe0925e7cb3d3b34ee9e856c4bc822a499975f)

Depends-On: Iee7c922cd534ff9ec67cffd2aaa1a03360c1b74a
Change-Id: Ic276cbbb91aaf73ba44e588a607338b93d056da8