XenAPI: Make cirros image file name configurable for XenServer
In stackrc, currently the cirros image file is hard coded as
cirros-0.3.4-x86_64-disk.vhd.tgz but in lib/tempest it sets
SCENARIO_IMAGE_FILE as
"cirros-${CIRROS_VERSION}-${CIRROS_ARCH}-disk.vhd.tgz".
Once CIRROS_VERSION or CIRROS_ARCH change, it will cause
mismatching. So let's change stackrc to depend on the variables
also. And change lib/tempest to use the variables defined in
stackrc. So we can control the overidable variables to specify
the cirros image version in a consistent way.

Change-Id: I6f48aa58dc5dcf5c18c48418d5ef4b4f2a6496eb