lib/neutron: configure root_helper for agents
Before the patch, we were only configuring root_helper_daemon to point
to oslo.rootwrap, but not root_helper. (The former is used for long
running commands only, while the latter is used for short lived
commands.) This made neutron agents to directly call to sudo when a
privileged process was to be executed. This failed because /etc/sudoers
was not configured to allow anything except the rootwrap call itself.

This patch simplifies rootwrap handling in the code; it also sets
root_helper to point to rootwrap; as well as configure daemon in
sudoers. Hopefully, it will be enough for agents to actually configure
backend as needed.

Change-Id: Ib05a6e0e024f534d7f616d41d70fb67ecf6daeaf