run_process: Add ability to specify a longer log name
Add the ability to specify a longer logfile name when processes are
not run with screen.

This will allow the ability to specify a more descriptive logfile
names in the gate to help newcomers to OpenStack more easily know what
the logfiles are when reading the filenames.

This will have no impact when running in a devstack environment using
screen. Thus the short names will still be used for the screen tabs.

Change-Id: I5f51a36c50191a8c6dcad2c7a80a022f19be58e6