Update patch set 5
Patch Set 5: Code-Review-1

I'm not really sure I understand this patch.

DevStack pretty much takes over the world that it has access to, and the point of this is to clean out any left over state so future runs aren't buggy. What other preserved content is in this NFS directory that you are giving DevStack full run of? Why isn't that just bind mounted at a lower point so that DevStack has a dedicated area to manage?

Patch-set: 5