Use qemu-kvm-ev package on centos
I9a972e3fde2e4e552f6fc98350820c07873c3de3 has bumped up the minimum
qemu version to 2.1.0, which is not provided by centos7 base.  This
has started to cause the centos7 job to fail

You have to use the qemu-kvm-ev package, which is based off the
qemu-kvm-rhev package, which is explained in [1] but you probably
can't read it.

The gist is, that qemu-kvm-rhev is a later build of kvm that is
incompatible with the base version provided.  qemu-kvm-rhev is only
provided with the RHV (ovirt) and RHOS (openstack) products.  CentOS
rebuilds this package as qemu-kvm-ev as part of it's virtualisation
SIG.

[1] https://access.redhat.com/solutions/629513
[2] https://wiki.centos.org/SpecialInterestGroup/Virtualization

Change-Id: Ide91b261f35fb19d8bd7155ca016fa3b76a45ea1