Update patch set 5
Patch Set 5:

The issue arises when running stack.sh on the compute nodes. By the time the compute node is ready for setup, the controller should have finished initializing. The compute node should not be attempting to delete the contents in the shared instance directories since the cleanup should be handled from the controller's end.

Patch-set: 5