Update patch set 1
Patch Set 1:

> This looks semantically fine. However, the cinder race bug seems to
 > be showing up in the full run - so maybe they're related? or maybe
 > not?

(b) maybe not.

A search on logstash for the cinder race bug that you referenced showed that it has happened a number of times. I think this is more like a practical example of Amrith's Law (Murphy was right, I'm the example, if something can go wrong, it will go wrong when I try it).

Patch-set: 1
Label: Workflow=0