Update patch set 3
Patch Set 3:

do they really represent the same thing?  if you add or remove something to the neutron-* files, it's not like you're really testing with the q-* deployment method too to ensure it's not broken in some odd way.

i wonder if it's better to just leave them separate and let them diverge if they must.

Patch-set: 3