Update patch set 5
Patch Set 5: Code-Review+1

this is interesting, comparing [1] (this run) to [2] (a peer run from recently) at peak memory usage times

   PID   %MEM             RSS       PPID       TIME     NLWP WCHAN                     COMMAND
 19592    3.5          284436          1   00:05:39      178 -                         /usr/sbin/mysqld
 19199    6.5          538928          1   00:02:04      149 -                         /usr/sbin/mysqld

both these peak memory usage points came about ~40 minutes into the run, so are roughly comparable.  ergo i agree roughly with the etherpad.

My only concern is this is really quite far from an experimental principle of holding everything constant and changing one thing.  Do we have a feeling for which of these ~20 settings is actually making the most difference here?  could we trim this down a bit so that we have less of a shotgun approach?  it might be better to figure this out now before the gate starts failing with mysql errors and we just have to revert the whole thing because there's too much pressure to do the selective debugging then.

but, i mean, it does seem to help reduce memory which is good

[1] http://logs.openstack.org/68/438668/5/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/a3230f9/logs/screen-peakmem_tracker.txt.gz
[2] http://logs.openstack.org/56/440956/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/41cf890/logs/screen-peakmem_tracker.txt.gz

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1