Update patch set 5
Patch Set 5:

> this is interesting, comparing [1] (this run) to [2] (a peer run
 > from recently) at peak memory usage times
 > 
 > PID   %MEM             RSS       PPID       TIME     NLWP WCHAN    
 >                 COMMAND
 > 19592    3.5          284436          1   00:05:39      178 -      
 >                   /usr/sbin/mysqld
 > 19199    6.5          538928          1   00:02:04      149 -      
 >                   /usr/sbin/mysqld
 > 
 > both these peak memory usage points came about ~40 minutes into the
 > run, so are roughly comparable.  ergo i agree roughly with the
 > etherpad.
 > 
 > My only concern is this is really quite far from an experimental
 > principle of holding everything constant and changing one thing. 
 > Do we have a feeling for which of these ~20 settings is actually
 > making the most difference here?  could we trim this down a bit so
 > that we have less of a shotgun approach?  it might be better to
 > figure this out now before the gate starts failing with mysql
 > errors and we just have to revert the whole thing because there's
 > too much pressure to do the selective debugging then.
 > 
 > but, i mean, it does seem to help reduce memory which is good
 > 
 > [1] http://logs.openstack.org/68/438668/5/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/a3230f9/logs/screen-peakmem_tracker.txt.gz
 > [2] http://logs.openstack.org/56/440956/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/41cf890/logs/screen-peakmem_tracker.txt.gz

Ian, thanks for the confirmation. I'd just like to point out that it isn't as if we're changing 20 things (we literally are, but that's not what I'm disputing). We are changing one thing which in MySQL is represented with twenty or so knobs. Each of the knobs contributes a bit. Of course, we could pick the biggest knob and change that first and then the next one, and so on.

My experience with adjusting MySQL memory has been based on very well documented 'here's how MySQL uses its memory' and therefore I feel quite comfortable with this kind of change, but I definitely appreciate the concern re: making so many changes at once (but I don't share them).

Thx!

Patch-set: 5