Make the _DEVSTACK_STACKRC variable a global
When using 'declare' it is by default a local variable. But the other
variables in stackrc are not using declare so they become globals and
they are exposed when a function sources stackrc to the caller of the
function. But _DEVSTACK_STACKRC was a local and not propagating like
the other variables.

Closes-Bug: #1669509
Change-Id: I193bd3e611a65672b5ace871358fd57ee5c5dbae