Make the sentry variables global
The sentry variables:
    _DEVSTACK_FUNCTIONS
    _DEVSTACK_FUNCTIONS_COMMON
    _DEVSTACK_STACKRC

were being declared with 'declare -r' to make them read-only. When
using 'declare' it is by default a local variable. But the other
variables are not using declare so they become global variables and
they are exposed when a function sources the file to the caller of the
function. But the sentry variables were local variables and not
propagating outside of a function that sources the file. This caused
errors to occur at times.

Closes-Bug: #1669509
Co-Authored-By: John L. Villalovos <john.l.villalovos@intel.com>
Change-Id: I2180b68fe861ad19c6d4ec0df0f9f8a528347862