Update patch set 2
Patch Set 2: Code-Review-1

I'm racking my brains to see why this is correct ... it's not clear to me sorry.

In [1] we have essentially

 $(source stackrc; echo $CONFIG_VAR)

but [2] we have DEVSTACK_PLUGINS set, but the gate var _DEVSTACK_STACKRC is *not* set.

this suggests to me that the plugin is being enabled outside the path of source stackrc->sources local.conf->enable_plugin?  If you follow that path I can't see how _DEVSTACK_STACKRC isn't set?  The only place we += to DEVSTACK_PLUGINS is in enable_plugin.

I feel like this might actually be telling us we've violated some other assumption?  maybe if this is set before sourcing we should be giving an error message that you're using plugins outside local.conf?  because it seems like this will silently override which isn't a great interface.  

or ... i might be wrong :)

[1] http://git.openstack.org/cgit/openstack-dev/grenade/tree/inc/upgrade#n28
[2] http://logs.openstack.org/57/436957/10/experimental/gate-grenade-dsvm-ironic-multinode-multitenant-ubuntu-xenial-nv/56b2faa/logs/grenade.sh.txt.gz#_2017-03-02_09_58_19_688

Patch-set: 2
Label: Code-Review=-1