Update patch set 2
Patch Set 2: Code-Review+2

Ok, so the tempest change defaults to false [1] and we want to turn it on for master using this.  We do *not* want to have to override this in the job definition of project-config [2]; those jobs run on every branch.  

However, the tempest review doesn't have positive votes yet, maybe because it has been caught up in how to configure it.  If tempest is happy with the job this looks fine

[1] https://review.openstack.org/#/c/433411/14/tempest/config.py
[2] https://review.openstack.org/#/c/438347/

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2