Update patch set 3
Patch Set 3: Code-Review-1

So this is causing errors with grenade as any variable that is done with "declare" does not exist outside of the function call. Which can been with the error with setting GITREPO["oslo.cache"] because it was declared as an associative array but without the -g.

So it is getting complicated to go this route :(

Patch-set: 3
Reviewer: Gerrit User 14760 <14760@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1