Update patch set 1
Patch Set 1: Workflow-1

OK, so I start to understand a bit better what's going on. Indeed, that patch wouldn't not fix it. I thought that pip_install would detect the source file, but it doesn't work that way.

The pip_install_gr in keystone correctly installs it with python3, as it's what's enabled. We could special case keystonemiddleware to install it in both versions if we detect that's necessary. Or have an additional call in lib/swift, and find a way to detect the source file.

Any other idea?

Patch-set: 1
Label: Workflow=-1