Append to logs within the same stack.sh run
Now if a plugin restarts some service, e.g. nova, previous
contents of the log are lost. The CURRENT_LOG_TIME is the same, so
no new log is created. This makes it a bit harder to debug problems that
might happen before a service was restarted.

From 'man bash':

  The format for appending standard output and standard error is:

         &>>word

  This is semantically equivalent to

         >>word 2>&1

Change-Id: I903d97ccbd68a6b6fe684b847bdacac88fa1c728