Set up sudoers to never require password for mlock_report.py
The tool is executed as part of memory_tracker service, and so it should
not require password. Even though we call the tool using a relative
path, we can still use an absolute path in sudoers entry because sudo
will expand the path for executed command before validating it against
sudoers rules.

Without the patch, memory_tracker service eventually dies when time
since last sudo execution is long enough for the cached sudo password to
expire. This usually happens after devstack has done its job, and
tempest tests are triggered.

Depends-On: If7926ce6a2996b766c49b010a7f6640ae624f860
Change-Id: I342ac402d431711c7f0ffaaf96f0b9a9a2e0990a