Update patch set 3
Patch Set 3: Code-Review-1

so part of the local.conf adds in d-g was so that we didn't have different ways to configure things like neutron in devstack from what's in the config file.

Is there a reason to not make things just specify the whole line in their local.conf snippet?

Patch-set: 3