Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

ok, so I want to step back a second. How much coverage is lost in Heat if we just keep swift off for the python3 porting running of Heat?

Because the only reason we were going to go this split route is that we assumed that all the infrastructure to run split route was really in there, but it's not as seen here. And while I don't want to hold up Heat getting some coverage here, I also think that the various brutal hacks we might need to put into devstack as we move forward are going to be pretty fragile and cause breaks on people when we remove them in 6 months time.

It would be good to sync with Doug (who is about to go on vacation) once he gets back to figure that out.

Patch-set: 2
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1