WIP: getting working systemd units
During the PTG there was a discussion that the screen developer
workflow wasn't nearly as useful as it once was. There were now too
many services to see them all on one screen, and one of the most
common service restart scenarios was not restarting one service, but a
bunch to get code to take effect.

Clark suggested user units under systemd. This would have the added
benefit of a converged code path between gate and development, as well
as getting folks more familiar with systemd workflow.

This is the first really basic stab here. Logging is ok, you can
access logs via journalctl as normal user (this really needs systemd
native integration from oslo.log to be ideal).

Change-Id: I836bf27c4cfdc449628aa7641fb96a5489d5d4e7