Update patch set 2
Patch Set 2:

(5 comments)

This is great. For the most part it lines up with how simple I expected this to be. Excited for the potential log goodness that journalctl would give us particularly are tracing requests or interleaving logs among a set of services (I know both of these are very common use cases).

Inline comments are minor things that I noticed to help clean it up a bit more but nothing crazy.

Patch-set: 2
Label: Verified=0