Update patch set 1
Patch Set 1:

We saw a few percent reduction in benchmark performance and no significant benefits for memory (the VMs are quite large and varied O/S levels). It may be a good default to turn it on but we would aim to turn it off on compute intensive hypervisors. For this kind of thing, we use puppet to decide whether to run it or not.

Patch-set: 1