Update patch set 1
Patch Set 1:

Luz, you are right. It seems it was still possible to install Devstack master on Trusty as of Feb 13th, so it makes sense to somehow not break support for Trusty just because of this.

Nevertheless, as it's just a matter of time before Devstack (especially Nova) won't work on Trusty, I will not abandon this, but just delay this patch.

Patch-set: 1
Label: Verified=0