Update patch set 11
Patch Set 11:

(1 comment)

Wow, like Clark said, that's a lot cleaner than I would have expected.  Nice work.

It doesn't look like this changes any of the screens that are running tail -f in a window, is it time to get rid of those?  Or did I miss that?

Patch-set: 11
Label: Verified=0