Update patch set 6
Patch Set 6:

(2 comments)

Looks ok. Incidentally, if you switch to using a layered config file, everything in that block is applicable to any service on any systemd host except the name of the service itself. The service name is libvirt-bin on Ubuntu, and libvirtd on Fedora/RHEL/CentOS.

Patch-set: 6
Label: Verified=0