Update patch set 7
Patch Set 7:

I rolled up my suggestions in this patch, and also incorporated the electric fence change:

https://review.openstack.org/451430

I've only tested it on Fedora 25, but I can confirm that it results in a libvirtd running with efence mapped and max core dump limit. If you want to roll any or all of it into your patch, feel free.

Patch-set: 7