Update patch set 1
Patch Set 1:

(1 comment)

@Jens ... we do install this as part of nova

@clarkb: doesn't seem to have taken though [1]? 

To me it's just a question of what people are expected to use in the wild.  It doesn't really make much sense to essentially skip libvirt 1.3.1 problems if people are trying to deploy on it.  but ... if people are really only deploying from UCA then this does make sense.  It's fairly similar to centos, where we don't use the base qemu packages because they're too old, but the qemu-ev packages which are done by RDO & RHOS (and centos ported by the centos virt SIG)

that said ... if we want to go this path, it might be worth doing this in the base images rather than here so that we get consistency in testing.  if you've got a bindep install, for example, seems like you'd get a different version for func testing than on devstack 

[1] http://logs.openstack.org/92/451492/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/9440843/logs/devstacklog.txt.gz#_2017-03-29_17_40_50_971

Patch-set: 1
Label: Verified=0