Update patch set 5
Patch Set 5:

Ah, another funny thing Ubuntu did, they changed the group name from "libvirtd" to "libvirt" after Xenial, in order to be consistent with Debian. Not sure how to best fix it, though. Probably create the new group as an alias to the old one, just need to decide when is the correct moment for that.

Patch-set: 5
Label: Verified=0