Update patch set 8
Patch Set 8: Code-Review-1

I think it's cool that this shows we *can* do this.  -1 just because I'd like to flag the *should* we do this as something needing maybe a little more than casual review.  I may be over-thinking the impact of this but I don't think I'd want to approve it until people more directly impacted have had a look.

For me, two things I'd like to be convinced on:

1) is this a good idea?  Since this is quite far ahead of base Xenial, and keeps getting further ahead at regular intervals, this seems to introduce a fairly large uncertainty window between base Xenial and Xenial+UCA.  Maybe nobody actually cares if things work on base Xenial because it's not relevant in the wild.  If you're on your own debugging things like this libvirt segfault on base Xenial, you'd have to say that for all practical purposes it has been abandoned.  Before one or two of us unilaterally make a decision on that, I would prefer to see some sort of consensus.

FWIW, centos uses RDO for a couple of things.  We use the updated qemu-ev packages which are RDO/RHOS but repackaged by the centos virt SIG team.   openvswitch is another.  of course the main problem is that the base qemu is not recent enough to run trunk nova.  but in the wild nobody is running without these packages, so it's a moot point.  

The other half of this is a gate-instability window around UCA releases.  Previously, updates of big things like this have just been when we introduced a new platform (trusty->xenial, say).  this seems to introduce more regular update points tied to new UCA releases, and potential for instability.

I'd be happy just seeing some people who would be affected by this (nova, i guess mostly) signing off.

2) is this the right way to do it?  if this is only enabled in devstack runs, then that seems to leave another window of uncertainty for not-devstack jobs.  functional testing for example, would be running against base versions.  is this a problem?

to use the centos example again, we have a default rdo install as part of stack.sh, but have a SKIP_EPEL_INSTALL that sees if we're on nodepool and knows that the repos are there from the build.  

I am feeling like it belongs in the base image for platform consistency, maybe with a similar skip to the RDO stuff as mentioned.  I would be happy to hear arguments why I'm wrong :)

Patch-set: 8
Label: Code-Review=-1