[WIP] Bring docker installation to devstack
There are several projects that tried to install docker in different
ways. This potentially lead to failure of more than one such services
were enabled. This commit consolidate docker installation into
devstack tree so that other services could depend on it.

The initial script was borrowed from Kuryr-libnetwork. If this
patch is accepted, Kuryr-libnetwork, Fuxi, and Zun intend to
leverage the common script to install Docker.

Change-Id: Icd3c2398681927801bfaa89eb89985f6751c6b31