Update patch set 1
Patch Set 1:

> My concern here is that nothing in default devstack uses this,
 > which means we're not going to be self testing this change. I would
 > rather dedicate a devstack-docker plugin (which could live under
 > the qa team), that also had +2 authority from key members of the
 > different container teams. So it could be fixed/enhanced by either
 > current QA team members or members of the container teams.

Hi Sean,

Personally, I don't mind moving the script to a dedicated devstack plugin, but I would like to mention that there are at least three projects that are committed to use it. You have my promise that Zun will leverage this script and I received a promise from Antoni that Kuryr-libnetwork and Fuxi will leverage it as well.

Patch-set: 1